Move site structure utilities to src/utils and add latest edition detection
**Reorganization:**
- Move site-structure.ts from scripts/src/ to src/utils/
- Move sidebar-from-site-structure.ts from scripts/src/ to src/utils/
- Update all imports in generate.ts and test files
- scripts/ is now leaf code that depends on utils, not vice versa

**Latest Edition Detection:**
- Add optional `folder` field to Edition interface
- Create `getLatestEditionName()` function that determines latest edition by:
  1. Edition with folder = "." (current directory)
  2. Edition named "general", "main", or "default"
  3. First edition in the list
- Update `addAEPEdition()` to accept optional folder parameter
- Update `assembleAEPItems()` to use `getLatestEditionName()`
- Set folder="." for general edition in generate.ts

**Code Formatting:**
- Clean up comments and formatting in sidebar assembly functions
- Improve readability of conditional checks

This shared logic can now be used by both sidebar assembly and other
components like EditionBanner that need to determine the latest edition.
